rename `GoToType` to `GoToTypeDef` (Go-to type definition)
